requirements.txt
httpx
pillow
coco-lib
tqdm
torch
torchvision
numpy
open_clip_torch
requests
pandas